backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
requests
google-genai
openai
tavily-python